use postgresql as db: change dtos

+++ src/main/java/it/mrt/bills/dtos/CommonParameterDTO.java
@@ -2,12 +2,14 @@ package it.mrt.bills.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CommonParameterDTO {
 
     private String name;
     private String category;
-    private Double value;
-    private Double start;
-    private Double end;
+    private BigDecimal value;
+    private Integer start;
+    private Integer end;
 }
