+++ src/main/java/it/mrt/bills/dtos/gas/GasComparisonRequestDTO.java
@@ -2,12 +2,13 @@ package it.mrt.bills.dtos.gas;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 public class GasComparisonRequestDTO {
 
-    private Double consumption;
+    private BigDecimal consumption;
     private String region;
 
     private String priceType;
